Always run 01 script in image building process
The CentOS images currently don't have the 01 script run in them. This
means that Docker images are pulled unnecessarily and the build time is
prolonged. This is because of a missing 'true' argument when the
provisioning script is run.

Previously, cloning metal3 and running the 01 script was gated in the
image building process by a variable called DEPLOY_METAL3, which took
the first argument the script was run with as a boolean. The
provisioning scripts explicitly state that they deploy metal3, so I
can't see a purpose for this variable and have removed it to ensure that
we consistently run the 01 script from metal3 in our images.