Refactor OpenStack volume creation
It should be possible to create a volume from an image in OpenStack
without booting any machines.

Every time a VM is booted in OpenStack from an image, the image is
copied to a volume. The volume is attached to a VM which is booted. We
do volume creation in the image build so that the the image to volume
copying step does not need to be repeated for every instance boot.

Since we no longer spin up a VM to do the volume creation step, the
cleanup_volume.sh script is no longer needed. (The packer image creation
does cleanup of VMs for the image building part of the pipeline on image
creation failure.)

I've kept the retry behavior from the old script, since this offers
protection from any CityCloud flakes.