Implement automated Harbor cleanup process
This commit:

Implements a general purpose harbor bash library.

Implements the cleanup process for harbor.

Moves the harbor cleanup process to the same script where the artifactory
cleanup process is handled and renames the file from artifactory_cleanup.sh to
artifact_cleanup.sh.

By using the aforementioned bash library implements a process to
save the digest of the ironic-image container image that is being
built during the IPA build process and saves the digest to the metadata file.