Fix cloud-init error on centos
resolv.conf was previously made immutable to prevent NetworkManager
and/or cloud-init from modifying it. This was causing issues with
cloud-init, since it still tried to modify it.

Most of the time things worked ok (except that cloud-init had error
status). But every now and then this error seemed to cause other issues.
E.g. the resizing of the filesystem or injection of authorized
ssh keys failed.

This commit properly configures NetworkManager to set the nameservers
and cloud-init to avoid touching resolv.conf.