Uplift IPA base image to centos 9 stream
This change was needed because centos 8 stream is slowly loosing upstream support and
the systemd packaged with centos 8 stream had a bug that affected udev thus the inspection
and provisioning process of Ironic.

This change also contains a additional element that puts a buildinfo file into IPA to
help with future debugging tasks.