Replace the old ssh key
The new ssh key was already added previously as a second key in all
images. This PR removes the old key from the image building pipeline and
adds the new in its place. After merging this, only the new key will be
available on the images.

This also removes the workaround for supporting legacy crypto policies,
since it is no longer needed when using the new key.