remove machine-id in ubuntu images
Currently the image contains an id in /etc/machine-id.
This file should contain a unique machine ID which
uniquely identifies the host (see machine-id(5)).
Some processes rely on this file to generate other
values e.g. mac-addresses on bond interfaces.
If the machine-id is the same accross many hosts
these mac-addresses will under certain circumstances
be generated identically. Two hosts having the same
mac address leads to L2-networking issues.

This commit empties the machine-id file so that it is
regenerated on the next boot (which will be the first
boot of the machine the image is used for).

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>